backend/app/services/research_service.py
import httpx
from typing import List, Dict, Any
from bs4 import BeautifulSoup
import json
import re


class ResearchService:
    """Service for conducting web research and keyword analysis"""

    def __init__(self):
        self.search_engines = []

    async def research_sector(
        self, sector: str, location: str, additional_keywords: str = None
    ) -> Dict[str, Any]:
        """
        Conduct comprehensive research on a sector in a specific location

        Args:
            sector: Industry sector (e.g., "Real Estate")
            location: Geographic location (e.g., "Ghana")
            additional_keywords: Optional additional search terms

        Returns:
            Dictionary containing research data and trending keywords
        """
        # Build search queries
        queries = self._build_search_queries(sector, location, additional_keywords)

        # Gather data from multiple sources
        trending_topics = []
        keywords = set()

        for query in queries:
            try:
                results = await self._search_web(query)
                trending_topics.extend(results.get("topics", []))
                keywords.update(results.get("keywords", []))
            except Exception as e:
                print(f"Error searching for '{query}': {e}")
                continue

        # Analyze and rank keywords
        ranked_keywords = self._rank_keywords(list(keywords), sector, location)

        # Extract trending topics
        top_topics = self._extract_top_topics(trending_topics, limit=5)

        return {
            "sector": sector,
            "location": location,
            "queries_used": queries,
            "trending_topics": top_topics,
            "keywords": ranked_keywords[:20],  # Top 20 keywords
            "search_volume_estimate": "high",  # Placeholder for actual search volume data
            "competition_level": "medium",  # Placeholder
        }

    def _build_search_queries(
        self, sector: str, location: str, additional_keywords: str = None
    ) -> List[str]:
        """Build effective search queries for research"""
        queries = [
            f"{sector} trends {location} 2024",
            f"latest {sector} news {location}",
            f"{sector} market {location}",
            f"popular {sector} topics {location}",
            f"{sector} insights {location}",
        ]

        if additional_keywords:
            keywords = [k.strip() for k in additional_keywords.split(",")]
            for keyword in keywords:
                queries.append(f"{keyword} {sector} {location}")

        return queries

    async def _search_web(self, query: str) -> Dict[str, Any]:
        """
        Perform web search using available APIs or scraping

        Note: In production, you would use Google Custom Search API,
        Bing Search API, or similar services. This is a simplified version.
        """
        topics = []
        keywords = set()

        try:
            # Simulate web search results
            # In production, replace with actual API calls
            async with httpx.AsyncClient() as client:
                # Example: Search for blog posts, news articles
                # For now, we'll use a simulated approach

                # Extract keywords from query
                query_keywords = self._extract_keywords_from_query(query)
                keywords.update(query_keywords)

                # Simulate finding trending topics
                topics.append({
                    "title": f"Latest trends in {query}",
                    "relevance": 0.9
                })

        except Exception as e:
            print(f"Search error: {e}")

        return {
            "topics": topics,
            "keywords": list(keywords)
        }

    def _extract_keywords_from_query(self, query: str) -> List[str]:
        """Extract potential keywords from search query"""
        # Remove common words
        stop_words = {
            "the", "a", "an", "in", "on", "at", "for", "to", "of", "and",
            "is", "are", "was", "were", "be", "been", "being", "2024", "2025"
        }

        words = re.findall(r'\b\w+\b', query.lower())
        keywords = [w for w in words if w not in stop_words and len(w) > 3]

        return keywords

    def _rank_keywords(
        self, keywords: List[str], sector: str, location: str
    ) -> List[str]:
        """Rank keywords by relevance and potential"""
        # Simple ranking based on length and uniqueness
        # In production, use actual search volume and competition data

        scored_keywords = []
        for keyword in keywords:
            score = 0

            # Longer keywords often more specific
            if len(keyword) > 8:
                score += 2

            # Keywords containing sector or location
            if sector.lower() in keyword.lower():
                score += 3
            if location.lower() in keyword.lower():
                score += 2

            scored_keywords.append((keyword, score))

        # Sort by score descending
        scored_keywords.sort(key=lambda x: x[1], reverse=True)

        return [kw for kw, score in scored_keywords]

    def _extract_top_topics(
        self, topics: List[Dict[str, Any]], limit: int = 5
    ) -> List[Dict[str, Any]]:
        """Extract and rank top trending topics"""
        # Sort by relevance if available
        sorted_topics = sorted(
            topics,
            key=lambda x: x.get("relevance", 0),
            reverse=True
        )

        return sorted_topics[:limit]

    def generate_blog_outline(
        self, research_data: Dict[str, Any], keywords: List[str]
    ) -> Dict[str, Any]:
        """
        Generate a blog outline based on research data

        Args:
            research_data: Research findings
            keywords: Target keywords

        Returns:
            Blog outline structure
        """
        sector = research_data.get("sector", "")
        location = research_data.get("location", "")
        trending_topics = research_data.get("trending_topics", [])

        outline = {
            "title_suggestions": [
                f"Top {len(trending_topics)} {sector} Trends in {location} for 2024",
                f"The Ultimate Guide to {sector} in {location}",
                f"What's Trending in {location}'s {sector} Market",
            ],
            "sections": [
                {
                    "heading": "Introduction",
                    "key_points": [
                        f"Overview of {sector} market in {location}",
                        "Why this matters now",
                        "What readers will learn"
                    ]
                },
                {
                    "heading": f"Current State of {sector} in {location}",
                    "key_points": [
                        "Market overview",
                        "Recent developments",
                        "Key statistics"
                    ]
                },
                {
                    "heading": "Trending Topics",
                    "key_points": [
                        topic.get("title", "") for topic in trending_topics[:3]
                    ]
                },
                {
                    "heading": "Key Insights and Opportunities",
                    "key_points": [
                        "What the data tells us",
                        "Emerging opportunities",
                        "Expert predictions"
                    ]
                },
                {
                    "heading": "Conclusion",
                    "key_points": [
                        "Summary of key findings",
                        "Actionable takeaways",
                        "Future outlook"
                    ]
                }
            ],
            "target_keywords": keywords[:5],
            "estimated_word_count": 1500,
        }

        return outline
