backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcryptbcrypt==4.1.2
python-dotenv==1.0.0

# Background tasks
celery==5.3.6
redis==5.0.1

# HTTP requests
httpx==0.26.0
requests==2.31.0

# Email
sendgrid==6.11.0

# AI/LLM
anthropic==0.18.1

# Payment processing
stripe==8.1.0
paystackapi==2.1.0

# PDF generation
reportlab==4.0.9
markdown2==2.4.12
weasyprint==60.2

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-cors==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-slugify==8.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
